+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added support for resolving 'ap' URIs with `gateways` parameter to `ap fetch` command.
+
 ## [5.7.1] - 2026-07-12
 
 ### Fixed