+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added support for resolving 'ap' URIs with `gateways` parameter to `ap fetch` and `ap import` commands.
+- Added support for resolving 'ap' URIs with `gateways` parameter to `/api/v2/search` API endpoint.
+- Accept portable `Delete(Person)` activities.
+
+## [5.7.1] - 2026-07-12
+
 ### Fixed
 
 - Changed application name returned by `--version` from `mitra_cli` to `mitra`.